Added the first version to make changes directly through the software.
The application is meant to provide customizability to users.
That is why we are introducing a way to directly changed important
settings like the text on the labels, stripe colors, to show a graph or
not and if to show then with what colors, etc. Try this our for your
own.